refactor: phrases module with standard structure
